moving qt4 directory inside new _ui_tester_register
