+++ traitsui/testing/tester/_ui_tester_register/qt4/_traitsui/check_list_editor.py
