+++ traitsui/testing/tester/_ui_tester_register/qt4/_traitsui/instance_editor.py
