+++ traitsui/testing/tester/_ui_tester_register/qt4/_traitsui/ui_base.py
@@ -10,7 +10,7 @@
 #
 from traitsui.qt4.ui_base import ButtonEditor
 from traitsui.testing.tester import command, query
-from traitsui.testing.tester.qt4 import _interaction_helpers
+from traitsui.testing.tester._ui_tester_register.qt4 import _interaction_helpers
 
 
 def register(registry):
