+++ traitsui/testing/tester/_ui_tester_register/qt4/default_registry.py
