+++ traitsui/testing/tester/_ui_tester_register/qt4/tests/test_interaction_helpers.py
