+++ traitsui/tests/editors/test_range_editor.py
@@ -29,7 +29,7 @@ def _register_simple_spin(registry):
     registry : TargetRegistry
         The registry being registered to.
     """
-    from traitsui.testing.tester.qt4 import _registry_helper
+    from traitsui.testing.tester._ui_tester_register.qt4 import _registry_helper
     from traitsui.qt4.range_editor import SimpleSpinEditor
 
     _registry_helper.register_editable_textbox_handlers(
